Move the movie method to resolve linters errors
